Merge gamma expansion and shader stages
For SolidColor and gradients, this means we can skip the expansion stage
for source by just doing it once to the solid color.  This also causes
gradients to be blended in the linear colorspace instead of using the
compressed values.

Pixmap patterns still apply gamma expansion after sampling because the
sampler integrates the load operations without the ability to insert a
gamma expansion pass.